Fixed Error raising and refactored duplicate Code into TransformerWithTargetMixin
