+++ category_encoders/leave_one_out.py
@@ -1,14 +1,14 @@
 """Leave one out coding"""
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.base import BaseEstimator
 import category_encoders.utils as util
 from sklearn.utils.random import check_random_state
 
 __author__ = 'hbghhy'
 
 
-class LeaveOneOutEncoder(BaseEstimator, TransformerMixin):
+class LeaveOneOutEncoder(BaseEstimator, util.TransformerWithTargetMixin):
     """Leave one out coding for categorical features.
 
     This is very similar to target encoding but excludes the current row's
@@ -209,20 +209,6 @@ class LeaveOneOutEncoder(BaseEstimator, TransformerMixin):
             return X
         else:
             return X.values
-
-    def fit_transform(self, X, y=None, **fit_params):
-        """
-        Encoders that utilize the target must make sure that the training data are transformed with:
-             transform(X, y)
-        and not with:
-            transform(X)
-        """
-
-        # the interface requires 'y=None' in the signature but we need 'y'
-        if y is None:
-            raise(TypeError, 'fit_transform() missing argument: ''y''')
-
-        return self.fit(X, y, **fit_params).transform(X, y)
 
     def fit_leave_one_out(self, X_in, y, cols=None):
         X = X_in.copy(deep=True)
