+++ category_encoders/m_estimate.py
@@ -1,7 +1,7 @@
 """M-probability estimate"""
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.base import BaseEstimator
 from category_encoders.ordinal import OrdinalEncoder
 import category_encoders.utils as util
 from sklearn.utils.random import check_random_state
@@ -9,7 +9,7 @@ from sklearn.utils.random import check_random_state
 __author__ = 'Jan Motl'
 
 
-class MEstimateEncoder(BaseEstimator, TransformerMixin):
+class MEstimateEncoder(BaseEstimator, util.TransformerWithTargetMixin):
     """M-probability estimate of likelihood.
 
     This is a simplified version of target encoder, which goes under names like m-probability estimate or
@@ -240,20 +240,6 @@ class MEstimateEncoder(BaseEstimator, TransformerMixin):
             return X
         else:
             return X.values
-
-    def fit_transform(self, X, y=None, **fit_params):
-        """
-        Encoders that utilize the target must make sure that the training data are transformed with:
-            transform(X, y)
-        and not with:
-            transform(X)
-        """
-
-        # the interface requires 'y=None' in the signature but we need 'y'
-        if y is None:
-            raise(TypeError, 'fit_transform() missing argument: ''y''')
-
-        return self.fit(X, y, **fit_params).transform(X, y)
 
     def _train(self, X, y):
         # Initialize the output
