+++ category_encoders/target_encoder.py
@@ -1,14 +1,14 @@
 """Target Encoder"""
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.base import BaseEstimator
 from category_encoders.ordinal import OrdinalEncoder
 import category_encoders.utils as util
 
 __author__ = 'chappers'
 
 
-class TargetEncoder(BaseEstimator, TransformerMixin):
+class TargetEncoder(BaseEstimator, util.TransformerWithTargetMixin):
     """Target encoding for categorical features.
 
     For the case of categorical target: features are replaced with a blend of posterior probability of the target
@@ -242,20 +242,6 @@ class TargetEncoder(BaseEstimator, TransformerMixin):
             return X
         else:
             return X.values
-
-    def fit_transform(self, X, y=None, **fit_params):
-        """
-        Encoders that utilize the target must make sure that the training data are transformed with:
-             transform(X, y)
-        and not with:
-            transform(X)
-        """
-
-        # the interface requires 'y=None' in the signature but we need 'y'
-        if y is None:
-            raise(TypeError, 'fit_transform() missing argument: ''y''')
-
-        return self.fit(X, y, **fit_params).transform(X, y)
 
     def target_encode(self, X_in):
         X = X_in.copy(deep=True)
