+++ category_encoders/woe.py
@@ -1,7 +1,7 @@
 """Weight of Evidence"""
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.base import BaseEstimator
 from category_encoders.ordinal import OrdinalEncoder
 import category_encoders.utils as util
 from sklearn.utils.random import check_random_state
@@ -9,7 +9,7 @@ from sklearn.utils.random import check_random_state
 __author__ = 'Jan Motl'
 
 
-class WOEEncoder(BaseEstimator, TransformerMixin):
+class WOEEncoder(BaseEstimator, util.TransformerWithTargetMixin):
     """Weight of Evidence coding for categorical features.
 
     Parameters
@@ -238,20 +238,6 @@ class WOEEncoder(BaseEstimator, TransformerMixin):
             return X
         else:
             return X.values
-
-    def fit_transform(self, X, y=None, **fit_params):
-        """
-        Encoders that utilize the target must make sure that the training data are transformed with:
-            transform(X, y)
-        and not with:
-            transform(X)
-        """
-
-        # the interface requires 'y=None' in the signature but we need 'y'
-        if y is None:
-            raise(TypeError, 'fit_transform() missing argument: ''y''')
-
-        return self.fit(X, y, **fit_params).transform(X, y)
 
     def _train(self, X, y):
         # Initialize the output
